+++ fabric/src/main/java/com/avp/fabric/data/compatibility/common/CommonConstants.java
@@ -51,6 +51,8 @@ public class CommonConstants {
 
     public static final TagKey<Item> RAW_MATERIALS_LEAD = createItem("raw_materials/lead");
 
+    public static final TagKey<Item> RAW_MATERIALS_TITANIUM = createItem("raw_materials/titanium");
+
     public static final TagKey<Item> URANIUM = createItem("ingots/uranium");
 
     /*